Refactoring to reuse code standard to htc json format
