+++ src/survive_playback.c
@@ -11,67 +11,14 @@
 //All MIT/x11 Licensed Code in this file may be relicensed freely under the GPL or LGPL licenses.
 
 #include <survive.h>
-#include <jsmn.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <errno.h>
+
 #include <string.h>
-#include <sys/stat.h>
-#include <os_generic.h>
-#if !defined(__FreeBSD__) && !defined(__APPLE__)
-#include <malloc.h> // for alloca
-#endif
 #include <sys/time.h>
-#include "json_helpers.h"
+
 #include "survive_config.h"
 #include "survive_default_devices.h"
-
-static int jsoneq(const char *json, jsmntok_t *tok, const char *s) {
- if (tok->type == JSMN_STRING && (int) strlen(s) == tok->end - tok->start &&
-    strncmp(json + tok->start, s, tok->end - tok->start) == 0) {
-		return 0;
-	}
-	return -1;
-}
-
-
-static int ParsePoints( SurviveContext * ctx, SurviveObject * so, char * ct0conf, FLT ** floats_out, jsmntok_t * t, int i )
-{
-	int k;
-	int pts = t[i+1].size;
-	jsmntok_t * tk;
-
-	so->nr_locations = 0;
-	*floats_out = malloc( sizeof( **floats_out ) * 32 * 3 );
-
-	for( k = 0; k < pts; k++ )
-	{
-		tk = &t[i+2+k*4];
-
-		int m;
-		for( m = 0; m < 3; m++ )
-		{
-			char ctt[128];
-
-			tk++;
-			int elemlen = tk->end - tk->start;
-
-			if( tk->type != 4 || elemlen > sizeof( ctt )-1 )
-			{
-				SV_ERROR( "Parse error in JSON\n" );
-				return 1;
-			}
-
-			memcpy( ctt, ct0conf + tk->start, elemlen );
-			ctt[elemlen] = 0;
-			FLT f = atof( ctt );
-			int id = so->nr_locations*3+m;
-			(*floats_out)[id] = f;
-		}
-		so->nr_locations++;
-	}
-	return 0;
-}
 
 struct SurvivePlaybackData {
   SurviveContext * ctx;
@@ -215,7 +162,7 @@ static int playback_poll( struct SurviveContext * ctx, void * _driver ) {
   return 0;
 }
 
-int playback_close( struct SurviveContext * ctx, void * _driver ) {
+static int playback_close( struct SurviveContext * ctx, void * _driver ) {
     SurvivePlaybackData* driver = _driver;
     if(driver->playback_file)
       fclose(driver->playback_file);
@@ -224,7 +171,7 @@ int playback_close( struct SurviveContext * ctx, void * _driver ) {
 }
 
 
-static int LoadConfig( SurvivePlaybackData * sv, SurviveObject * so, int devno, int iface, int extra_magic )
+static int LoadConfig( SurvivePlaybackData * sv, SurviveObject * so)
 {
 	SurviveContext * ctx = sv->ctx;
 	char * ct0conf = 0;
@@ -241,91 +188,12 @@ static int LoadConfig( SurvivePlaybackData * sv, SurviveObject * so, int devno, 
 	fseek(f, 0, SEEK_SET);  //same as rewind(f);
 
 	ct0conf = malloc(len+1);       
-	fread( ct0conf, len, 1, f);
+	int read = fread( ct0conf, len, 1, f);
 	fclose( f );
 	ct0conf[len] = 0;
 
 	printf( "Loading config: %d\n", len );
-
-	if (len == 0)
-	  return 1;
-
-		//From JSMN example.
-		jsmn_parser p;
-		jsmntok_t t[4096];
-		jsmn_init(&p);
-		int i;
-		int r = jsmn_parse(&p, ct0conf, len, t, sizeof(t)/sizeof(t[0]));	
-		if (r < 0) {
-			SV_INFO("Failed to parse JSON in HMD configuration: %d\n", r);
-			return -1;
-		}
-		if (r < 1 || t[0].type != JSMN_OBJECT) {
-			SV_INFO("Object expected in HMD configuration\n");
-			return -2;
-		}
-
-		for (i = 1; i < r; i++) {
-			jsmntok_t * tk = &t[i];
-
-			char ctxo[100];
-			int ilen = tk->end - tk->start;
-			if( ilen > 99 ) ilen = 99;
-			memcpy(ctxo, ct0conf + tk->start, ilen);
-			ctxo[ilen] = 0;
-
-//				printf( "%d / %d / %d / %d %s %d\n", tk->type, tk->start, tk->end, tk->size, ctxo, jsoneq(ct0conf, &t[i], "modelPoints") );
-//				printf( "%.*s\n", ilen, ct0conf + tk->start );
-
-			if (jsoneq(ct0conf, tk, "modelPoints") == 0) {
-				if( ParsePoints( ctx, so, ct0conf, &so->sensor_locations, t, i  ) )
-				{
-					break;
-				}
-			}
-			if (jsoneq(ct0conf, tk, "modelNormals") == 0) {
-				if( ParsePoints( ctx, so, ct0conf, &so->sensor_normals, t, i  ) )
-				{
-					break;
-				}
-			}
-
-
-			if (jsoneq(ct0conf, tk, "acc_bias") == 0) {
-				int32_t count = (tk+1)->size;
-				FLT* values = NULL;
-				if ( parse_float_array(ct0conf, tk+2, &values, count) >0 ) {
-					so->acc_bias = values;
-					so->acc_bias[0] *= .125; //XXX Wat?  Observed by CNL.  Biasing by more than this seems to hose things.
-					so->acc_bias[1] *= .125;
-					so->acc_bias[2] *= .125;
-				}
-			}
-			if (jsoneq(ct0conf, tk, "acc_scale") == 0) {
-				int32_t count = (tk+1)->size;
-				FLT* values = NULL;
-				if ( parse_float_array(ct0conf, tk+2, &values, count) >0 ) {
-					so->acc_scale = values;
-				}
-			}
-
-			if (jsoneq(ct0conf, tk, "gyro_bias") == 0) {
-				int32_t count = (tk+1)->size;
-				FLT* values = NULL;
-				if ( parse_float_array(ct0conf, tk+2, &values, count) >0 ) {
-					so->gyro_bias = values;
-				}
-			}
-			if (jsoneq(ct0conf, tk, "gyro_scale") == 0) {
-				int32_t count = (tk+1)->size;
-				FLT* values = NULL;
-				if ( parse_float_array(ct0conf, tk+2, &values, count) >0 ) {
-					so->gyro_scale = values;
-				}
-			}
-		}
-
-	return 0;
+	return survive_load_htc_config_format(ct0conf, len, so);
 }
 
 
@@ -357,11 +225,11 @@ int DriverRegPlayback( SurviveContext * ctx )
   SurviveObject * tr0 = survive_create_tr0(ctx, "Playback", sp);
   SurviveObject * ww0 = survive_create_ww0(ctx, "Playback", sp);
   
-  if( !LoadConfig( sp, hmd, 1, 0, 0 )) { survive_add_object( ctx, hmd ); }
-  if( !LoadConfig( sp, wm0, 2, 0, 1 )) { survive_add_object( ctx, wm0 ); }
-  if( !LoadConfig( sp, wm1, 3, 0, 1 )) { survive_add_object( ctx, wm1 ); }
-  if( !LoadConfig( sp, tr0, 4, 0, 0 )) { survive_add_object( ctx, tr0 ); }
-  if( !LoadConfig( sp, ww0, 5, 0, 0 )) { survive_add_object( ctx, ww0 ); }
+  if( !LoadConfig( sp, hmd )) { survive_add_object( ctx, hmd ); }
+  if( !LoadConfig( sp, wm0 )) { survive_add_object( ctx, wm0 ); }
+  if( !LoadConfig( sp, wm1 )) { survive_add_object( ctx, wm1 ); }
+  if( !LoadConfig( sp, tr0 )) { survive_add_object( ctx, tr0 ); }
+  if( !LoadConfig( sp, ww0 )) { survive_add_object( ctx, ww0 ); }
 
   
   survive_add_driver(ctx, sp, playback_poll, playback_close, 0);
